refactor: moved some code from media.component.ts to post.component.ts
