+++ client/projects/post/src/app/post/media/media.component.ts
@@ -1,8 +1,6 @@
-import {Component, Inject, Input, OnInit} from '@angular/core';
+import {Component, EventEmitter, Inject, Input, OnInit, Output} from '@angular/core';
 import {Post} from "../../model/post";
 import {MAT_DIALOG_DATA, MatDialog} from "@angular/material/dialog";
-import {GQLService} from "../../service/gql.service";
-import {AuthLibService} from "auth-lib";
 
 @Component({
   selector: 'app-post-media',
@@ -12,58 +10,22 @@ import {AuthLibService} from "auth-lib";
 export class MediaComponent implements OnInit {
 
   @Input() post: Post;
+  @Output() newEvent = new EventEmitter<string>();
   editMode: boolean = false;
 
-  constructor(private dialog: MatDialog,
-              private gqlService: GQLService,
-              private authService: AuthLibService) { }
+  constructor(private dialog: MatDialog) { }
 
   ngOnInit(): void {
   }
 
-  likePost(post: Post) {
-    // TODO uncomment this
-    const username: string = /*this.authService.userName*/ 'user4';
-    let liked: boolean = true;
-
-    // check if it's a like or unlike
-    const index = post.likedBy.indexOf(username, 0);
-    if (index > -1) {
-      const newLikedBy: string[] = []
-      post.likedBy.forEach((user, i) => {
-        if (index !== i) {
-          newLikedBy.push(user);
-        }
-      });
-      post.likedBy = newLikedBy;
-      liked = false;
-    } else {
-      const newLikedBy: string[] = [username];
-      post.likedBy.forEach(user => newLikedBy.push(user));
-      post.likedBy = newLikedBy;
-    }
-
-    this.gqlService.likePost(post, liked);
-  }
-
-  editPost(post: Post) {
-    post.editMode = false;
-    this.gqlService.editPost(post);
-  }
-
-  removePost(post: Post) {
-    post.editMode = false;
-    this.gqlService.removePost(post);
+  triggerEvent(eventName: string) {
+    this.newEvent.emit(eventName);
   }
 
   openLikedByDialog(likedBy: string[]) {
     this.dialog.open(DialogLikedBy, {
       data: likedBy
     });
-  }
-
-  showComments(post: Post) {
-    this.gqlService.getPostComments(post);
   }
 }
